+++ lib/models/listings/shared/options.js
@@ -1,0 +1,4 @@
+exports.options = {
+    timestamps: true,
+    discriminatorKey: 'kind'
+};
