+++ lib/route/listing.js
@@ -1,11 +1,12 @@
 // load Schema
-var Listings = require('../models/listing.js');
+var listingHandler = require('../handlers/listing.js');
 
 // Tusk Libraries
 var Validate = require('../validation.js');
 var Utils = require('../utils.js');
 var Consts = require('../consts.js');
 var Common = require('../common.js');
+
 var User = require('../models/user.js')
 var email = require('../email.js')
 
@@ -49,105 +50,8 @@ function updateMiscListingLatLng(listing, callback) {
     }
 }
 
-function updateSubletListingLatLng(listing, callback) {
-    if (listing.kind != Listings.SUBLET || typeof(listing.apt_info.address) === typeof(void 0)) {
-        return callback(false);
-    } else {
-        geocoder.geocode(listing.apt_info.address, function(err, res) {
-            if (!Array.isArray(res)) {
-                return callback(false);
-            }
-            switch(res.length) {
-                case 0:
-                    return callback(false);
-                case 1:
-                    var lat = res[0].latitude;
-                    var lng = res[0].longitude;
-                    listing.apt_info.lat = lat;
-                    listing.apt_info.lng = lng;
-                    return callback(true);
-                default:
-                    console.log("Unexpected number of geocode results for "+new_address);
-                    console.log(res);
-                    return callback(false);
-            }
-        });
-    }
-}
 
-function makeNewListingFromPost(body, user_id, callback) {
-    body.user_id = user_id;
-    var paths = [];
-    if ('type' in body) {
-        if (body.type === Listings.MISC) {
-            var miscListing = new Listings.MiscListing;
-            for (var path in Listings._MiscSchema.paths) {
-                var first_period = path.indexOf('.');
-                var real_path = path;
-                if (first_period != -1) {
-                    real_path = path.substring(0, first_period);
-                }
-                if (paths.indexOf(real_path) == -1) {
-                    paths.push(real_path);
-                }
-            }
-            Utils.copyObjectData(miscListing, body, paths);
-            return updateMiscListingLatLng(miscListing, function(result) {
-                if (result) {
-                    return callback({
-                        type: body.type,
-                        listing: miscListing
-                    });
-                } else {
-                    return callback(false);
-                }
-            });
-        } else if (body.type === Listings.SUBLET) {
-            var subletListing = new Listings.SubletListing;
-            for (var path in Listings._SubletSchema.paths) {
-                var first_period = path.indexOf('.');
-                var real_path = path;
-                if (first_period != -1) {
-                    real_path = path.substring(0, first_period);
-                }
-                if (paths.indexOf(real_path) == -1) {
-                    paths.push(real_path);
-                }
-            }
-            Utils.copyObjectData(subletListing, body, paths);
-            return updateSubletListingLatLng(subletListing, function(result) {
-                if (result) {
-                    callback({
-                        type: body.type,
-                        listing: subletListing
-                    });
-                } else {
-                    return callback(false);
-                }
-            });
-        } else if (body.type === Listings.BOOK) {
-            var bookListing = new Listings.BookListing;
-            for (var path in Listings._BookSchema.paths) {
-                var first_period = path.indexOf('.');
-                var real_path = path;
-                if (first_period != -1) {
-                    real_path = path.substring(0, first_period);
-                }
-                if (paths.indexOf(real_path) == -1) {
-                    paths.push(real_path);
-                }
-            }
-            Utils.copyObjectData(bookListing, body, paths);
-            return callback({
-                type: body.type,
-                listing: bookListing
-            });
-        } else if (book.type === Listings.FURNITURE) {
 
-        }
-    }
-    return callback(false);
-}
 
 exports.postListing = function(request, response) {
     response.set('Content-Type', 'application/json');
@@ -244,6 +148,8 @@ exports.postListingById = function(request, response) {
 function secureListings(request, listings, include_id) {
     return listings.map(function(e) {
         return secureListing(request, e, include_id);
+    }).filter(function(e) {
+        //
     });
 }
 
@@ -254,6 +160,7 @@ function secureListing(request, listing, include_id) {
         } else
             return listing.toSecure(Consts.ROLE_INVALID, include_id);
     } else if (Common.ensureLoginSession(request)) {
+        //TODO: warn and log this case... should NEVER happen!
         if (Consts.checkPriv(request.session.who.role, Consts.ROLE_ADMIN)) {
             return listing;
         } else if (Consts.checkPriv(request.session.who.role, Consts.ROLE_MODERATOR_PUBLIC)
@@ -456,3 +363,17 @@ exports.putListingAttributeById = function(request, response) {
         }
     });
 }
+
+exports.putListingVisibilityById = function(request, response) {
+    response.set('Content-Type', 'application/json');
+    if (!Common.ensureLoginSession(request)) {
+        return response.status(400).send(Common.error(
+            'NOT_LOGGED_IN_EXCEPTION',
+            'Listing could not be posted because the user is not logged in'
+        ));
+    }
+    var uid = request.params.listing_id;
+    var visible = request.params.visible;
+
+}
+
